requirements.txt
APScheduler==3.7.0
asyncio==3.4.3
certifi==2021.5.30
chardet==4.0.0
fake-useragent==0.1.11
idna==3.2
pytz==2021.1
pywin32==228
requests==2.25.1
selenium==3.141.0
six==1.16.0
tzlocal==2.1
urllib3==1.26.4
win32-setctime==1.0.3
wxPython==4.1.0
pypubsub==4.0.3
